sfw/scripts/requirements.txt
Pillow == 8.3.1
argparse == 1.4.0
numpy == 1.21.1
pyserial == 3.5
async-timeout == 4.0.2
certifi == 2022.12.7
charset-normalizer == 2.1.1
idna == 3.4
redis == 4.4.0
requests == 2.28.1
setuptools == 47.1.0
six == 1.16.0
spotipy == 2.21.0
urllib3 == 1.26.13